+++ lib/karafka-core.rb
@@ -48,3 +48,8 @@ rd_config = ::Rdkafka::Config
 rd_config.statistics_callback = instrumentation.statistics_callbacks
 rd_config.error_callback = instrumentation.error_callbacks
 rd_config.oauthbearer_token_refresh_callback = instrumentation.oauthbearer_token_refresh_callbacks
+
+# This loads librdkafka components into memory prior to initializing the client.
+# This mitigates macos forking issues.
+# @see https://github.com/confluentinc/librdkafka/issues/4590
+::Rdkafka::Bindings.rd_kafka_global_init