+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Karafka Core Changelog
 
 ## 2.4.6 (Unreleased)
+- [Enhancement] Mitigate macos forking issues when librdkafka is not loaded to memory.
 - [Change] Allow `karafka-rdkafka` `0.18.0`.
 
 ## 2.4.5 (2024-11-19)